streamlit_sources/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements.in
#
altair==5.4.1
    # via streamlit
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
bbcode==1.1.0
    # via -r requirements.in
blinker==1.8.2
    # via streamlit
cachetools==5.5.0
    # via
    #   google-auth
    #   streamlit
certifi==2024.8.30
    # via requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via streamlit
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via streamlit
google-api-core[grpc]==2.23.0
    # via
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-storage
google-auth==2.36.0
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-firestore
    #   google-cloud-storage
google-cloud-core==2.4.1
    # via
    #   google-cloud-firestore
    #   google-cloud-storage
google-cloud-firestore==2.18.0
    # via -r requirements.in
google-cloud-storage==2.18.0
    # via -r requirements.in
google-crc32c==1.5.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos==1.66.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.68.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.68.0
    # via google-api-core
idna==3.10
    # via requests
importlib-resources==6.4.5
    # via
    #   jsonschema
    #   jsonschema-specifications
jinja2==3.1.4
    # via
    #   altair
    #   pydeck
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2023.12.1
    # via jsonschema
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
narwhals==1.14.2
    # via altair
numpy==1.24.4
    # via
    #   pandas
    #   pyarrow
    #   pydeck
    #   streamlit
packaging==24.2
    # via
    #   altair
    #   streamlit
pandas==2.0.3
    # via streamlit
pillow==10.4.0
    # via streamlit
pkgutil-resolve-name==1.3.10
    # via jsonschema
proto-plus==1.25.0
    # via
    #   google-api-core
    #   google-cloud-firestore
protobuf==5.28.3
    # via
    #   google-api-core
    #   google-cloud-firestore
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
    #   streamlit
pyarrow==17.0.0
    # via streamlit
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pydeck==0.9.1
    # via streamlit
pygments==2.18.0
    # via rich
python-dateutil==2.9.0.post0
    # via pandas
pytz==2024.2
    # via pandas
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   google-api-core
    #   google-cloud-storage
    #   streamlit
rich==13.9.4
    # via streamlit
rpds-py==0.20.1
    # via
    #   jsonschema
    #   referencing
rsa==4.9
    # via google-auth
six==1.16.0
    # via python-dateutil
smmap==5.0.1
    # via gitdb
streamlit==1.40.1
    # via -r requirements.in
tenacity==9.0.0
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.4.2
    # via streamlit
typing-extensions==4.12.2
    # via
    #   altair
    #   rich
    #   streamlit
tzdata==2024.2
    # via pandas
urllib3==2.2.3
    # via requests
zipp==3.20.2
    # via importlib-resources
